Added NCBITaxa support to eggnog-mapper
Up to this commit, eggnog-mapper depended on the NCBITaxa module
from ete3 to perform taxa filtering and to print species names.
Now, we adapted the module and embed it directly within eggnog-mapper
so that no ete3 dependency is required to use eggnog-mapper.
Also, we included the required sqlite3 database and traverse .pkl file,
and appropiate files for testing.